requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
async-timeout==4.0.2
asyncio-throttle==1.0.2
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
powerline-gitstatus==1.3.1
powerline-shell==0.7.0
powerline-status===2.8.2.dev9999-git.b-82c1373ba0b424c57e8c12cb5f6f1a7ee3829c27-
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pygame==2.1.0
pymongo==4.0.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
tqdm==4.62.3
twitter==1.19.3
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
